# api/openapi.yaml
openapi: 3.0.1
info:
  title: Time-addressable Media Store
  description: |
    The Time-addressable Media Store (TAMS) is used for storing segmented media flows.
    The TAMS consists of a media store for the media flow segment objects and a service providing
    a database index of the flow segments. This document is a specification of the service API.

    See the [bbc/tams](https://github.com/bbc/tams) repository for more background on TAMS.
  version: '5.1'
paths:
...
  /flows:
    head:
      summary: List Flows
      description: Return flows path headers
      operationId: HEAD_flows
      tags:
      - Flows
      parameters:
      - name: source_id
        in: query
        description: Filter on source identifier.
        schema:
          $ref: '#/components/schemas/uuid'
      - name: timerange
        in: query
        description: Filter on flows that overlap the given timerange.
        schema:
          default: _
          $ref: '#/components/schemas/timerange'
      - name: format
        in: query
        description: Filter on flow format.
        schema:
          $ref: '#/components/schemas/contentformat'
      - name: codec
        in: query
        description: Filter on flow codec.
        schema:
          $ref: '#/components/schemas/mimetype'
      - name: label
        in: query
        description: Filter on flows that have the given label.
        schema:
          type: string
      - name: tag.name #{name}
        in: query
        description: |
          Filter on flows that have a tag named {name} and with the given value.
          The {name} could contain escaped characters to allow it to be used in a
          URL.
        schema:
          type: string
      - name: tag_exists.name #{name}
        in: query
        description: |
          Filter on flows that have a tag named {name} regardless of value. The
          {name} could contain escaped characters to allow it to be used in a
          URL. If set to true then the presence of the tag is filtered for. If set
          to false then its absence is. If left out then no filtering on tag presence
          is performed.
        schema:
          type: boolean
      - name: frame_width
        in: query
        description: Filter on video flows that have the given frame width.
        schema:
          type: integer
      - name: frame_height
        in: query
        description: Filter on video flows that have the given frame height.
        schema:
          type: integer
      - $ref: '#/components/parameters/trait_resource_paged_key'
      - $ref: '#/components/parameters/trait_paged_limit'
      responses:
        '200':
          description: ''
          headers:
            Link:
              description: Provides references to cursors for paging. Only the 'rel' attribute with value 'next' and a link to the next page is currently supported. If 'next' is not present then it is the last page.
              schema:
                type: string
            X-Paging-Limit:
              description: Identifies the current limit being used for paging. This may not match the requested value if the requested value was too high for the implementation
              schema:
                type: integer
            X-Paging-NextKey:
              description: Opaque string that can be supplied to the `page` query parameter to get the next page of results.
              schema:
                type: string
          content:
            application/json:
              schema:
                type: string
        '400':
          description: Bad request. Invalid query options.
      security:
      - Cognito:
        - openid
        - aws.cognito.signin.user.admin
        - tams-api/read
      x-amazon-apigateway-request-validator: parameters-only
      x-amazon-apigateway-integration:
        type: aws_proxy
        httpMethod: POST
        uri:
          Fn::Sub: arn:${AWS::Partition}:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/${FlowsFunction.Arn}/invocations
        passthroughBehavior: when_no_templates
    get:
      summary: List Flows
      description: List the flows registered in the store.
      operationId: GET_flows
      tags:
      - Flows
      parameters:
      - name: source_id
        in: query
        description: Filter on source identifier.
        schema:
          $ref: '#/components/schemas/uuid'
      - name: timerange
        in: query
        description: Filter on flows that overlap the given timerange.
        schema:
          default: _
          $ref: '#/components/schemas/timerange'
      - name: format
        in: query
        description: Filter on flow format.
        schema:
          $ref: '#/components/schemas/contentformat'
      - name: codec
        in: query
        description: Filter on flow codec.
        schema:
          $ref: '#/components/schemas/mimetype'
      - name: label
        in: query
        description: Filter on flows that have the given label.
        schema:
          type: string
      - name: tag.name #{name}
        in: query
        description: |
          Filter on flows that have a tag named {name} and with the given value.
          The {name} could contain escaped characters to allow it to be used in a
          URL.
        schema:
          type: string
      - name: tag_exists.name #{name}
        in: query
        description: |
          Filter on flows that have a tag named {name} regardless of value. The
          {name} could contain escaped characters to allow it to be used in a
          URL. If set to true then the presence of the tag is filtered for. If set
          to false then its absence is. If left out then no filtering on tag presence
          is performed.
        schema:
          type: boolean
      - name: frame_width
        in: query
        description: Filter on video flows that have the given frame width.
        schema:
          type: integer
      - name: frame_height
        in: query
        description: Filter on video flows that have the given frame height.
        schema:
          type: integer
      - $ref: '#/components/parameters/trait_resource_paged_key'
      - $ref: '#/components/parameters/trait_paged_limit'
      responses:
        '200':
          description: ''
          headers:
            Link:
              description: Provides references to cursors for paging. Only the 'rel' attribute with value 'next' and a link to the next page is currently supported. If 'next' is not present then it is the last page.
              schema:
                type: string
            X-Paging-Limit:
              description: Identifies the current limit being used for paging. This may not match the requested value if the requested value was too high for the implementation
              schema:
                type: integer
            X-Paging-NextKey:
              description: Opaque string that can be supplied to the `page` query parameter to get the next page of results.
              schema:
                type: string
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/flow'
        '400':
          description: Bad request. Invalid query options.
      security:
      - Cognito:
        - openid
        - aws.cognito.signin.user.admin
        - tams-api/read
      x-amazon-apigateway-request-validator: parameters-only
      x-amazon-apigateway-integration:
        type: aws_proxy
        httpMethod: POST
        uri:
          Fn::Sub: arn:${AWS::Partition}:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/${FlowsFunction.Arn}/invocations
        passthroughBehavior: when_no_templates
...
components:
  schemas:
    uuid:
      title: UUID
      pattern: ^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$
      type: string
    contentformat:
      title: Content Format
      description: Identifies the content format for a flow or source using a URN string.
      enum:
      - urn:x-nmos:format:video
      - urn:x-nmos:format:audio
      - urn:x-nmos:format:data
      - urn:x-nmos:format:multi
      type: string
    mimetype:
      title: MIME Type
      pattern: .*/.*
      description: MIME Type string.
      type: string
...
    flow:
      type: object
      description: Describes a Flow
      title: Flow
      oneOf:
      - $ref: '#/components/schemas/flowvideo'
      - $ref: '#/components/schemas/flowaudio'
      - $ref: '#/components/schemas/flowdata'
      - $ref: '#/components/schemas/flowmulti'
...
    timerange:
      title: TimeRange
      description: |
        A timerange of timestamps. It is represented using a pair ot timestamps with inclusivity and exclusivity markers.

        E.g.
        * "[0:0_10:0)" represents 10 seconds of media starting at timestamp "0:0" and ending before "10:0".
        * "(5:0_" represents a timerange starting after "5:0" and to eternity.
        * "[1694429247:0_1694429248:0)" is a 1 second TAI timerange starting at 2023-09-11T10:46:50.0Z UTC.
        * A "[" or "]" indicates that bound is inclusive, and a "(" or ")" indicates that bound is exclusive.

        Details of the format can be found in the TimeRange class provided in the Python
        [mediatimestamp](https://github.com/bbc/rd-apmm-python-lib-mediatimestamp) library.
      type: string
      pattern: ^(\[|\()?(-?\d+:\d+)?(_(-?\d+:\d+)?)?(\]|\))?$
...
    flowaudio:
      type: object
      description: Describes an audio Flow
      title: Audio Flow
      allOf:
      - $ref: '#/components/schemas/flowcore'
      - type: object
        required:
        - format
        - essence_parameters
        - codec
        properties:
          format:
            description: The primary content type URN for the flow.
            type: string
            enum:
            - urn:x-nmos:format:audio
          essence_parameters:
            type: object
            description: Describes the parameters of the essence inside this audio Flow
            title: Audio Flow Essence Parameters
            required:
            - sample_rate
            - channels
            additionalProperties: false
            properties:
              sample_rate:
                description: The fixed number of samples per second.
                type: integer
                exclusiveMinimum: 0
              channels:
                description: The channel count.
                type: integer
                exclusiveMinimum: 0
              bit_depth:
                description: The number of significant bits used to represent the audio sample. The minumum number of bytes then equals `round_up(bit_depth / 8)`. If codec is `audio/x-raw-int` bit_depth must be set. If codec is `audio/x-raw-float` bit_depth must be set to 32 or 64
                type: integer
                exclusiveMinimum: 0
              codec_parameters:
                title: Audio Codec Parameters
                type: object
                required: []
                properties:
                  coded_frame_size:
                    description: The fixed number of samples per coded audio frame.
                    type: integer
                  mp4_oti:
                    description: The MPEG-4 Object Type Identification. For more information on the use of this property in codec strings, see https://developer.mozilla.org/en-US/docs/Web/Media/Formats/codecs_parameter#mpeg-4_audio
                    type: integer
              unc_parameters:
                title: Uncompressed Audio Parameters
                type: object
                required:
                - unc_type
                properties:
                  unc_type:
                    description: The uncompressed audio multi-channel representation type. If codec is `audio/x-raw-int` or `audio/x-raw-float`, unc_type must be set.
                    type: string
                    enum:
                    - interleaved
                    - planar
                    - pairs
    flowdata:
      type: object
      description: Describes a data Flow
      title: Data Flow
      allOf:
      - $ref: '#/components/schemas/flowcore'
      - type: object
        required:
        - format
        - essence_parameters
        - codec
        properties:
          format:
            description: The primary content type URN for the flow.
            type: string
            enum:
            - urn:x-nmos:format:data
          essence_parameters:
            type: object
            description: Describes the parameters of the essence inside this data Flow
            title: Data Flow Essence Parameters
            additionalProperties: false
            properties:
              data_type:
                description: The type of information encoded in the flow, identified using a URN. e.g. The data_type may be urn:x-tams:data:bounding-box, and the codec `application/json`.
                type: string
    flowmulti:
      type: object
      description: Describes a multi-essence Flow
      title: Multi-essence Flow
      allOf:
      - $ref: '#/components/schemas/flowcore'
      - type: object
        required:
        - format
        properties:
          format:
            description: The primary content type URN for the flow.
            type: string
            enum:
            - urn:x-nmos:format:multi
    flowvideo:
      type: object
      description: Describes a video Flow
      title: Video Flow
      allOf:
      - $ref: '#/components/schemas/flowcore'
      - type: object
        required:
        - format
        - essence_parameters
        - codec
        properties:
          format:
            description: The primary content type URN for the flow.
            type: string
            enum:
            - urn:x-nmos:format:video
          essence_parameters:
            type: object
            description: Describes the parameters of the essence inside this video Flow
            title: Video Flow Essence Parameters
            required:
            - frame_width
            - frame_height
            additionalProperties: false
            properties:
              frame_rate:
                description: The fixed number of frames per second. If this parameter is unset, the frame_rate is either unknown or variable.
                type: object
                required:
                - numerator
                properties:
                  numerator:
                    description: numerator
                    type: integer
                    exclusiveMinimum: 0
                  denominator:
                    description: denominator
                    type: integer
                    default: 1
                    exclusiveMinimum: 0
              frame_width:
                description: The width of the picture in pixels.
                type: integer
                exclusiveMinimum: 0
              frame_height:
                description: The height of the picture in pixels.
                type: integer
                exclusiveMinimum: 0
              bit_depth:
                description: The number of significant bits used to represent the video component sample. If codec is `video/raw`, bit_depth must be set.
                type: integer
                exclusiveMinimum: 0
              interlace_mode:
                description: Interlaced video mode for frames in this Flow
                type: string
                enum:
                - progressive
                - interlaced_tff
                - interlaced_bff
                - interlaced_psf
              colorspace:
                description: Colorspace used for the video
                type: string
                enum:
                - BT601
                - BT709
                - BT2020
                - BT2100
              transfer_characteristic:
                description: Transfer characteristic
                type: string
                enum:
                - SDR
                - HLG
                - PQ
              aspect_ratio:
                description: The display aspect ratio. i.e. display_width / display_height
                type: object
                required:
                - numerator
                - denominator
                properties:
                  numerator:
                    description: numerator
                    type: integer
                    exclusiveMinimum: 0
                  denominator:
                    description: denominator
                    type: integer
                    exclusiveMinimum: 0
              pixel_aspect_ratio:
                description: The pixel aspect ratio. This is usually 1:1 (i.e. square pixels) for modern video. Some, usually older, video formats use non-square pixels e.g. some Standard Definition video. This is where that may be indicated.
                type: object
                required:
                - numerator
                - denominator
                properties:
                  numerator:
                    description: numerator
                    type: integer
                    exclusiveMinimum: 0
                  denominator:
                    description: denominator
                    type: integer
                    exclusiveMinimum: 0
              component_type:
                description: Picture component representation.
                type: string
                enum:
                - YCbCr
                - RGB
              horiz_chroma_subs:
                description: Horizontal chroma component sub-sampling. When unc_type is set to a YUV type, horiz_chroma_subs must be set.
                type: integer
                exclusiveMinimum: 0
              vert_chroma_subs:
                description: Vertical chroma component sub-sampling. When unc_type is set to a YUV type, vert_chroma_subs must be set.
                type: integer
                exclusiveMinimum: 0
              unc_parameters:
                type: object
                title: Uncompressed Video Parameters
                required:
                - unc_type
                properties:
                  unc_type:
                    description: Uncompressed picture packing type. If codec is `video/raw`, unc_type must be set.
                    type: string
                    enum:
                    - planar
                    - YUYV
                    - UYVY
                    - AYUV
                    - v210
                    - v216
                    - RGB
                    - RGBx
                    - xRGB
                    - BGRx
                    - xBGR
                    - RGBA
                    - ARGB
                    - BGRA
                    - ABGR
                    - alpha
              avc_parameters:
                type: object
                title: AVC Codec Parameters
                required:
                - profile
                - level
                - flags
                properties:
                  profile:
                    description: AVC / H.264 profile byte. For more information on the use of this property in codec strings, see https://developer.mozilla.org/en-US/docs/Web/Media/Formats/codecs_parameter#using_the_codecs_parameter
                    type: integer
                  level:
                    description: AVC / H.264 level byte. For more information on the use of this property in codec strings, see https://developer.mozilla.org/en-US/docs/Web/Media/Formats/codecs_parameter#using_the_codecs_parameter
                    type: integer
                  flags:
                    description: AVC / H.264 flags byte. For more information on the use of this property in codec strings, see https://developer.mozilla.org/en-US/docs/Web/Media/Formats/codecs_parameter#using_the_codecs_parameter
                    type: integer
...
  parameters:
    trait_paged_limit:
      name: limit
      in: query
      description: Restrict the response to the specified number of results. Implementations may specify their own default and maximum for the limit
      schema:
        type: integer
    trait_resource_paged_key:
      name: page
      in: query
      description: Opaque string used by backend to access a specific page of results. Clients should read the next URL from the `Link` header returned with responses, or use value of the returned X-Paging-NextKey header. If not supplied, the first page is accessed. Implementations should ensure a consistent sort order is applied to pages of results.
      schema:
        type: string
  securitySchemes:
    Cognito:
      in: header
      name: Authorization
      type: apiKey
      x-amazon-apigateway-authorizer:
        providerARNs:
        - Fn::Sub: arn:${AWS::Partition}:cognito-idp:${AWS::Region}:${AWS::AccountId}:userpool/${UserPool}
        type: cognito_user_pools
      x-amazon-apigateway-authtype: cognito_user_pools
...